requirements_gpbo.txt
joblib
gpytorch
botorch
Cython
matplotlib==3.5.1
matplotlib-inline==0.1.6
networkx==2.6.3
numpy==1.21.2
pathtools==0.1.2
prompt-toolkit==3.0.38
PyTDC==0.4.0
PyYAML==6.0
ray==2.4.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
tqdm
wandb